+++ lib/dlms-cosem/include/dlms/cosem/simple_objects.hpp
@@ -877,6 +877,43 @@ private:
   CosemAccessRights rights_;
 };
 
+class CosemSpecialDaysTableObject : public ICosemObject
+{
+public:
+  static const std::uint8_t MaxSupportedVersion = 0u;
+
+  CosemSpecialDaysTableObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& entries,
+    AttributeAccessMode entriesAccess);
+  CosemSpecialDaysTableObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& entries,
+    AttributeAccessMode entriesAccess,
+    std::uint8_t version);
+
+  CosemObjectDescriptor Descriptor() const;
+  CosemAccessRights AccessRights() const;
+  CosemStatus ReadAttribute(
+    std::uint8_t attributeId,
+    CosemByteBuffer& output) const;
+  CosemStatus WriteAttribute(
+    std::uint8_t attributeId,
+    const CosemByteBuffer& input);
+  CosemStatus InvokeMethod(
+    std::uint8_t methodId,
+    const CosemByteBuffer& input,
+    CosemByteBuffer& output);
+
+  const CosemByteBuffer& Entries() const;
+  void SetEntries(const CosemByteBuffer& value);
+
+private:
+  CosemObjectDescriptor descriptor_;
+  CosemByteBuffer entries_;
+  CosemAccessRights rights_;
+};
+
 enum class CosemClockBase
 {
   NotDefined = 0,